Use MUnit in FrameEncoderSuite
